+++ generator/page_generators/documentation/panther/intrinsics/builtin_module_options.js
@@ -1,0 +1,44 @@
+//////////////////////////////////////////////////////////////////////////////////////////////////
+//                                                                                              //
+// Part of pcit-project.github.io, under the Apache License v2.0 with LLVM and PCIT exceptions. //
+// You may not use this file except in compliance with the License.                             //
+// See `https://github.com/PCIT-Project/PCIT-CPP/blob/main/LICENSE`for info.                    //
+//                                                                                              //
+//////////////////////////////////////////////////////////////////////////////////////////////////
+
+
+
+const Page = require("../../../../Page.js");
+const html = require("../../../../html.js");
+const terms = require("../../../../terms.js");
+const search = require("../../../../search.js");
+const syntax_highlighting = require("../../../../syntax_highlighting/syntax_highlighting.js");
+
+
+
+exports.getPageGenerator = function(){
+	return new (require("../../../../PageGenerator.js").PageGenerator)(
+		() => {
+			let page = new Page.Page(__filename, {
+				path        : "documentation/panther/intrinsics/builtin_module_options.html",
+				title       : "Builtin Module @options",
+				categories  : [search.Category.PANTHER, search.Category.DOCUMENTATION],
+				breadcrumbs : [Page.Breadcrumbs.DOCUMENTATION, Page.Breadcrumbs.PANTHER_DOCUMENTATION, Page.Breadcrumbs.PANTHER_INTRINSICS],
+				description : "Documentation for the builtin module @options in the Panther programming language",
+			});
+
+			// syntax_highlighting.addPantherIntrinsicType("@options", page, "");
+
+			return page;
+		},
+		(page) => {
+
+			page.paragraph(`The builtin module ${page.inlineCode("@options")} is for accessing module options. The module options are accessed as member variables.`);
+
+
+			page.h2("Example");
+			page.text(html.italic("(TODO)"));
+		}
+	);
+}
+